feature/productionize: WIP
Unfortunately, it doesn't look like using names volumes and
remapping the ids is going to work with the current setup.

Creating directories and files from the shell script will create
them with the permissions of the local nginxcertbot user and thus
the containers won't have r/w permissons on the files or dirs.